+++ apis/study_metadata/study_location.py
@@ -2,6 +2,7 @@
 from flask_restx import Resource, fields
 from flask import request
 from model import Study, db, StudyLocation
+from ..authentication import is_granted
 
 
 from apis.study_metadata_namespace import api
@@ -42,12 +43,11 @@ class StudyLocationResource(Resource):
 
     def post(self, study_id: int):
         """Create study location metadata"""
-        data = request.json
-
         study_obj = Study.query.get(study_id)
-
+        if not is_granted("study_metadata", study_obj):
+            return "Access denied, you can not delete study", 403
+        data = request.json
         list_of_elements = []
-
         for i in data:
             if "id" in i and i["id"]:
                 study_location_ = StudyLocation.query.get(i["id"])
@@ -69,6 +69,9 @@ class StudyLocationUpdate(Resource):
 
     def delete(self, study_id: int, location_id: int):
         """Delete study location metadata"""
+        study_obj = Study.query.get(study_id)
+        if not is_granted("study_metadata", study_obj):
+            return "Access denied, you can not delete study", 403
         study_location_ = StudyLocation.query.get(location_id)
 
         db.session.delete(study_location_)
